refactor: componentize pay reducer
