+++ renderer/reducers/pay/constants.js
@@ -1,0 +1,9 @@
+export const QUERY_FEES = 'QUERY_FEES'
+export const QUERY_FEES_SUCCESS = 'QUERY_FEES_SUCCESS'
+export const QUERY_FEES_FAILURE = 'QUERY_FEES_FAILURE'
+export const QUERY_ROUTES = 'QUERY_ROUTES'
+export const QUERY_ROUTES_SUCCESS = 'QUERY_ROUTES_SUCCESS'
+export const QUERY_ROUTES_FAILURE = 'QUERY_ROUTES_FAILURE'
+export const SET_REDIRECT_PAY_REQ = 'SET_REDIRECT_PAY_REQ'
+export const SET_REDIRECT_LN_URL = 'SET_REDIRECT_LN_URL'
+export const DECLINE_LNURL_WITHDRAWAL = 'DECLINE_LNURL_WITHDRAWAL'
