+++ renderer/reducers/pay/index.js
@@ -1,0 +1,7 @@
+import payReducer from './reducer'
+
+export default payReducer
+export paySelectors from './selectors'
+export * from './constants'
+export * from './reducer'
+export * from './ipc'
