+++ renderer/reducers/pay/messages.js
@@ -1,0 +1,8 @@
+import { defineMessages } from 'react-intl'
+
+// use pattern [reducer]_message
+
+/* eslint-disable max-len */
+export default defineMessages({
+  pay_lnurl_withdraw_error: 'Unable to process withdraw request from {service}: {reason}',
+})
